feat: allow multiple options for install,
    * Update README.md to reflect below changes
    * add bash function to get machine architecture
    * function matches case output options to existing release assets
    * use if,elif,else block to allow script to work with wget or curl

I noted that uname -m on arm is aarch64 but release assets are named arm64,
the added function matches the case to what is set in release assets build.

Also, the instruction in [README.md](https://github.com/techsolx/kustomize-sops#install-the-latest-release)
reference the use of curl and the script calls for wget, so I allowed
both.

I noted that the use of `source <(input)` will cause shell to
reset or close silently, if the script fails.

This silent failure will also close any tmux session, if the user is a tmux users.

I opted to pipe the output to bash as this script requires it anyway.

I note some similar type work here:
https://github.com/viaduct-ai/kustomize-sops/pull/139/

Successful test with darwin x86_64 and linux x86_64 and arm64.